+++ CHANGELOG.md
@@ -13,8 +13,12 @@
 - Remove button role for links with an href (see: https://github.com/AdeleD/react-paginate/pull/390)
 - Add a rel attribute on previous page, current page and next page controls. They are respectivement customizable thanks to props `prevPageRel`, `selectedPageRel` and `nextPageRel`. To disable the rel attribute, set them to `null`. (see: https://github.com/AdeleD/react-paginate/pull/391)
 - A new prop `disabledLinkClassName` has been added. It allows to add a custom class on disabled `previous` and `next` controls. Default to `undefined` (no class added).
+- `tabindex` is now set to `-1` on disabled controls (disable previous/next buttons; active page button). (see: https://github.com/AdeleD/react-paginate/issues/219)
 - Add a warning when providing an `initialPage` prop value that is greater than the maximum page index from `pageCount` prop.
 - Add a warning when providing a `forcePage` prop value that is greater than the maximum page index from `pageCount` prop.
+- The `href` generated from `hrefBuilder` prop is now also set active page control `<a>` tag. (see #242)
+- A new prop `hrefAllControls` has been added. It allows to enable `hrefBuilder` for all controls.
+- Fix page range of 2 when first page is selected. (see https://github.com/AdeleD/react-paginate/issues/307)
 
 ## >= 7.1.5